renamed detailed item view component
