+++ src/containers/App/index.js
@@ -19,7 +19,7 @@ import CategoryList from '../CategoryList';
 import LoginPage from '../LoginPage';
 import RegistrationPage from '../RegistrationPage';
 import NotFound from '../../components/NotFound';
-import EditItem from '../EditItem';
+import DetailedItem from '../DetailedItem';
 
 class App extends Component {
   
@@ -36,7 +36,7 @@ class App extends Component {
           <Switch>
             <Route exact path="/" component={HomePage} />
             <Route path="/category/:category" component={CategoryList} />
-            <Route exact path="/items/:id" component={EditItem}/>
+            <Route exact path="/items/:id" component={DetailedItem}/>
             <Route exact path='/login' component={LoginPage} />
             <Route exact path='/register' component={RegistrationPage} />
             <Route component={NotFound} />
