+++ src/containers/DetailedItem/index.js
